Skeleton Physics
This is a basic implementation of the physics handling for the user's
skeleton. The SkeletonBody object contains all of the necessary
SkeletonPoints for a user's skeleton. A new SkeletonBody object is
created for each user. Each SkeletonPoint corresponds to a certain part
of the user's body (head,neck, left elbow, etc.) and actively tracks
that points position and speed at a specified polling period. It also
can output to a .csv for graphing and stuff. Let me know what you
guysthink.